+++ Toucan/Views/Components/ModeSelectorBar.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace Toucan.Views;
+
+public partial class ModeSelectorBar : UserControl
+{
+    public ModeSelectorBar()
+    {
+        InitializeComponent();
+    }
+}