+++ Toucan/Views/Components/ZenEditorView.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace Toucan.Views;
+
+public partial class ZenEditorView : UserControl
+{
+    public ZenEditorView()
+    {
+        InitializeComponent();
+    }
+}